Separate projections errors handling from projections
