+++ delta/app/src/main/scala/ch/epfl/bluebrain/nexus/delta/wiring/StreamModule.scala
@@ -5,7 +5,7 @@ import ch.epfl.bluebrain.nexus.delta.sdk.ResourceShifts
 import ch.epfl.bluebrain.nexus.delta.sdk.projects.ProjectContextCache
 import ch.epfl.bluebrain.nexus.delta.sdk.stream.GraphResourceStream
 import ch.epfl.bluebrain.nexus.delta.sourcing.config.{ProjectionConfig, QueryConfig}
-import ch.epfl.bluebrain.nexus.delta.sourcing.projections.Projections
+import ch.epfl.bluebrain.nexus.delta.sourcing.projections.{ProjectionErrors, Projections}
 import ch.epfl.bluebrain.nexus.delta.sourcing.stream._
 import ch.epfl.bluebrain.nexus.delta.sourcing.stream.pipes._
 import ch.epfl.bluebrain.nexus.delta.sourcing.{DeleteExpired, PurgeElemFailures, Transactors}
@@ -47,8 +47,13 @@ object StreamModule extends ModuleDef {
     Projections(xas, cfg.query, cfg.restartTtl)
   }
 
-  make[Supervisor].fromResource { (projections: Projections, cfg: ProjectionConfig) =>
-    Supervisor(projections, cfg)
+  make[ProjectionErrors].from { (xas: Transactors, cfg: ProjectionConfig) =>
+    ProjectionErrors(xas, cfg.query)
+  }
+
+  make[Supervisor].fromResource {
+    (projections: Projections, projectionErrors: ProjectionErrors, cfg: ProjectionConfig) =>
+      Supervisor(projections, projectionErrors, cfg)
   }
 
   make[DeleteExpired].fromEffect {
