+++ delta/sourcing-psql/src/main/resources/scripts/postgres/init/V1_09_M03_001__elem_errors_index.ddl
@@ -1,0 +1,4 @@
+-----------------------------------------------------
+-- Index on the instant column of failed_elem_logs --
+-----------------------------------------------------
+CREATE INDEX IF NOT EXISTS failed_elem_logs_instant_idx ON public.failed_elem_logs(instant);
